Use more lenient BeautifulSoup parser
Maniphest Tasks: T130383

Summary:
Python's built-in parser errors unconditionally on SVG files that have
an incorrect encoding definition, regardless of whether or not they
can still be rendered.

Refs T130383

Fix variable name collision

Thumbor's base engine already uses an icc_profile variable which collides with this one (in the form of it being set to None instead of being unset).

Differential Revision: https://phabricator.wikimedia.org/D157