Work around Tornado default 100MB remote file limit
We could increase the limit, but the buffer in its
entirety would be stored in memory, which would
start causing issues for very large files (possibly
going over thumb's memory limit just to load the
file).

This should make things more efficient, as for most
engines we were writing the contents of the buffer
to disk anyway.

Also:
 - make huge source tests use remote files
 - make engines use only command line when possible
    in order to leverage subprocess limits

Refs T145785
Refs T145768