Add STL engine
Summary: This adds an engine for STL files, which runs them through 3d2png. Detection of STL files is accomplished by checking the entire file size as compared to the triangle count in the header. In the case of a Swift loader, however, we short-circuit that detection by simply adding 'solid' to the beginning of the comment section of the header.

Reviewers: Gilles, fgiunchedi

Reviewed By: Gilles

Subscribers: fgiunchedi, jenkins

Differential Revision: https://phabricator.wikimedia.org/D732