engine.vips: fall back to tinyrgb when libpng complains
libvips uses libpng to write out png files. If the input file contains
an ICC profile that libpng considers invalid, libpng will refuse to
write the png. A libvips developer indicated that a fix for this would
have to come from libpng, not libvips. Since this is documented,
intentional behavior, that seems unlikely.

When libvips thumbnail generation fails with "profile" in the error
message, this patch forces it to convert the thumbnail to TinyRGB
before writing it. The error handling code was changed to a for loop
for DRY and because a request for an out-of-bounds page with a bad ICC
profile would need three tries (input error, output error, result).

While changing the color profile of an image is undesireable, I am of
the opinion that any thumbnail is better than no thumbnail. The best
solution would be to bypass libpng altogether and write to a lossless
webp. However, the Stretch version of libvips strips all metadata from
webp outputs, which would mean that the image gets forced to use
TinyRGB anyway (with all the other normally-retained metadata removed).
When libvips is running 8.6.4+, this option should be considered.
Another possible solution would be to remove the ICC profile, store it,
and then re-add it later. The imagemagick engine already does this, but
I felt that trying to re-implement that functionality in this engine
would introduce unnecessary complexity.

One test case has been added. I had difficulty getting the webp test to
pass, so a closer look at that would be a good idea.

Upstream: https://github.com/libvips/libvips/issues/1635
Bug: T219569
Change-Id: I4ad0a88152c5b65b9808bb7b1040a1ceb4bfee85