engine.imagemagick: Catch error when pyexiv2 can't find metadata
A backported security fix in exiv2 0.25-3.1+deb9u1 sometimes causes
images with no metadata to be raised as a corrupt metadata error.
Since the only field we use from pyexiv2 is the optional orientation, it
is safe to continue without the metadata from pyexiv2 when it can't be
loaded in this case. Pyexiv2 uses a RuntimeError for any error it
doesn't know, plus a few others, so I decided to only catch the error
when the message matched.

This patch also ensures that unhandled exceptions in metadata.read()
are correctly logged. They were previously not logged, which made
tracking down this bug more difficult.

Bug: T245440
Change-Id: I02f8a32c66fb46dd6a300916cb9848466a746093