engine.vips: Don't run vips if scaling_factor == 1
When scaling_factor == 1, vips shrink does not shrink the image.
However, it still processes the image, which can cause undesired effects
(stricter profile interpretation, larger file size). Instead of running
vips and hoping it does nothing, skip running it entirely.

While investigating this bug, I discovered that vips PNG tests do not
automatically set a filetype of PNG, and BaseWikimediaEngine assumes JPG
in that case. This should not be a problem in production, as the Swift
handler sets the filetype. To make sure the tests function correctly,
the filetype is explicitly set to png. This caused the test_png image to
be slightly larger.

Bug: T218272
Change-Id: I478d2f2b20eeedd11e7c405fbca1589bbd797d49