Remove division operation hack related to Python2
* It was found that there are some places in the code
where float() and .0 are used to avoid Python2
behaviour related to dividing integers. In Python2
dividing two integers returns an integer.
Now the project is under Python3 and such hack is
unnecessary. Because Python3 returns not only
integers after dividing integers but also float
numbers. The result depends on numerical data.
As a result the hack was removed.
* math.floor() was removed where it can be
replaced with floor division - //

Bug: T314393
Depends-On: I5cd557c2ab25076e190886dc92bb51c07101d349
Change-Id: I580f5a1fa524ca878b2648e7f6dbac09927bb9f6