Provide additional tests to cover errors caused by wrong engine commands
* Pytest monkeypatch is used to replace a real path of an executable with
a fake one. It helps to return the stderr value from Popen.communicate()
with an error inside. @pytest.mark.usefixtures is used before each test
to apply such behaviour only to one test, it makes us sure that others
will be executed as expected.
* IndexError exception can be skipped by code coverage in stl.py since
it will not be raised. It was added just in case, struct.unpack() always
returns a tuple with at least one parameter.
* Since inject_fictures() function is pretty much the same in each test,
I tried to move inject_fictures() function to integration/__init__.py.
But there are some incompatibilities between pytest and unittest related
to parameterizing of tests and I also need to pass the fake path to the
fixture for each test, so it seems hard for now.

Bug: T318406
Change-Id: Ia54d71c2723197bddcf281a08dc990cbd0202069