Add tinyrgb colour profile
Currently we distribute this colour profile via puppet to metal
instances. With the move to k8s we have three options, include this in
the helm chart, package it as its own apt package or include it in
this repo. I have a preference for #2 long term, but for the short
term of the migration I would like to include this here. File
introduced in T100976.

Bug: T323775
Change-Id: I9e1e1606eb1b67d71c49dd940ac5f648741b0c1c