Use blubber via Docker tooling; no longer requires local binary
This puts the proper magic header on the blubber.yaml so that
Docker can fetch blubbler as a Docker image and use it to build
the images without a local blubber installation.

Confirmed that I can now run `make docker_test` and
`make docker_code-coverage`, and start up the main build/dev server
on macOS with Docker Desktop.

Change-Id: Ib4007f10b1af9b83d6107812bbce91c8fb6eb296