Update Thumbor repository according to the latest changes
- Remove the .arclint and .arcconfig files since the plan
to move from Gerrit to Phab Differential was scrapped.
- Make the launch of all existing Thumbor tests and flake8
linter during the build step of the deployment pipeline,
previously only offline tests were run.
- Update the docker-compose.yaml file according to the new
working flow of the dev project version. There is no need
to create Docker volume to run the development version
of the project.
- Review the Makefile and update its commands. Make the
linter run in the make test command. Remove test_3d.py
from the online tests group since it doesn't make any
requests to third-party services and it doesn't require an
internet connection.
- Update README.md according to the current state of the
project.
- Fix typos in pytest.ini.
- Remove setup.py since Wikimedia thumbor isn't used as a
package on PyPi or somewhere else to distribute. It is only
used within one environment which is created by a Dockerfile.
Moreover, all blubber.yaml file variants run successfully
and we haven't stumbled upon any issues related to not using
setup.py.
- Delete the duplicate tinyrgb.icc file. Change the path to
the tinyrgb.icc file in the tests to the new one that is in
the project's root directory.
- Fix linter errors.
- Remove usage of tox since the project uses Docker.

Bug: T325811
Change-Id: Ib5751334c6a05430b14abf0f99740805d6cc5520