imagemagick: use JSON output from exiftool
According to the author of exiftool, output encoding is not guaranteed
when using the human-readable output. Given that we're already parsing
the output we shouldn't be using this output method and should use
JSON, but doubly so when we might be endangered by unknown or invalid
user input in EXIF tags. Now that we explicitly decode this string as
utf-8, we need to enforce unicode output (exiftool's default).

Bug: T327887
Change-Id: Ifa80c335dcb3f6932ebc3c90b16f5d19f85a3c2b