imagemagick: rename self.exif to self.exif_dict
In BaseEngine, self.exif has long been used for storage of exif data -
but only as a binary object, not as a dict as we do. We never really
hit upon this collision and ensuing TypeErrors because we never
invoked the paths in BaseEngine that used this value. However since
the bump to 7.3.10, load() now interacts with self.exif and we're
consistently seeing errors where a slice is taken of our dict, leading
to errors in prod logs. These errors are fairly short on detail also
so they are just noise, even if they are not causing explicit failures
in image processing.

In future we should evaluate if it makes sense to do our own exif
processing at all

Bug: T331995
Change-Id: I12afeb8718b607a02c7d33e48c661cc724199942