svg: use rsvg-convert output flag
In some versions of rsvg-convert, error output does not go to stderr
and is included in stdout. This is a bug, but in general we leave
ourselves vulnerable to this kind of thing by relying on command
output when tools themselves have support for writing out to files. We
should avoid using stdout in general where possible. Write out to
tempfiles as an intermediate step, then clean up no matter what
happens.

Bug: T334725
Change-Id: I04f12730ed28c2ca54b725c6f4abdaf2b43c7f47