imagemagick: update test cases for fixes within libraries
A type change in imagemagick has changed the expected output for these
images. Going forward all images will be of a similar size. Currently
tests fail with this latest upstream version as the generated files
are larger than our acceptable degree of variance - a result of the
slightly larger used in imagemagick's libraries.

Pin dependencies for the short-term in advance of us removing our own
patched version of the libraries in question.

Bug: T334863
Change-Id: I84ffabb2b3a46a58be4847a58d8f995baafdefff