svg: attempt to build valid locales from hyphenated languages
Currently when we receive a language like "zh-hk" we currently fail to
serve the language in the SVG correctly. Currently we attempt to get
the correct image by altering LC_ALL, which works fine for languages
like "zh" but despite having a "zh_hk" language we will fail for
"zh-hk".

To attempt to remedy this, try to convert to a valid locale. If we
don't get a valid locale, try to get something valid by stripping off
parameters after the first -. This isn't ideal but in some cases will
get us a valid locale from a RFC 5646-style language string which we
can't handle at present (for example "az-latn" is a valid RFC 5646
language, but it isn't a valid locale).

Add a test to verify that we generate correct images when the lang
parameter is set.

Bug: T337139
Change-Id: I7ffac71c9c1170a742777d1475e7c2ce9a4f172f