images: enable "debug" on memcache, log when servers are dead
It seems like thumbor instances are never successfully talking to
memcache. Annoyingly, the memcache library will silently fail if it
can't talk to a server and then mark it unhealthy. Furthermore any
attempts to read or write to memcached will silently fail if a server
is unhealthy. Add some debug logging to show that, and enable the
"debug" log for the memcache client which will give us basic error
messages on a failure to connect.

Bug: T341805
Change-Id: I6b746f47e9f1969f5c9013a1c5bb557ee6446706