Don't ignore imagemagick exit status
Revert f321adbfa13dd1dcca41e10e55b7e57b57087a3d. T179200 was fixed
another way, so this workaround is hopefully no longer needed.

In T344233, the -resize operator fails but -extent succeeds, so instead
of a resized thumbnail, we get the original image cropped to the size of
a thumbnail, taking the center part of it. That seems like a perverse
failure mode.

Bug: T344233
Change-Id: I30758ae4c495a859cea04e2370969226caed6269