imagemagick: ignore all py3exiv2 exceptions
py3exiv2 is only used to extract Orientation information, so we don't
need to care about any errors accessing other parts of the file. It's
better to let exiftool and ImageMagick try to get the rest of the
metadata and thumbnail the file since they're both more tolerant of
errors. Any legitimate problems will be caught immediately after when
exiftool reads the file.

The T178072 mitigation appears to be no longer necessary with the switch
to py3exiv2.

Bug: T381594
Change-Id: I7113773453b50351b5133fb15e5786f9250b859e